Entity split into smaller directory
