+++ src/main/java/entity/item/Equipment.java
@@ -1,4 +1,4 @@
-package entity.equipment_slots.item;
+package entity.item;
 
 public abstract class Equipment extends Item implements Comparable<Equipment> {
     private final String statType; // the type of stat
