+++ src/main/java/entity/item/Weapon.java
@@ -1,4 +1,4 @@
-package entity.equipment_slots.item;
+package entity.item;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
