+++ src/main/java/entity/monster/Steal.java
@@ -1,7 +1,7 @@
 package entity.monster;
 
 import entity.Character;
-import entity.equipment_slots.item.Collectible;
+import entity.item.Collectible;
 import entity.player.Player;
 import java.util.Random;
