+++ src/main/java/file_reader/deserialization/BasicEquipmentSlotsDeserialization.java
@@ -6,9 +6,9 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import entity.equipment_slots.item.Armor;
+import entity.item.Armor;
 import entity.equipment_slots.BasicEquipmentSlots;
-import entity.equipment_slots.item.Weapon;
+import entity.item.Weapon;
 
 import java.io.IOException;
