+++ src/main/java/file_reader/deserialization/CollectibleInventoryDeserialization.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
-import entity.equipment_slots.item.Collectible;
+import entity.item.Collectible;
 import entity.inventory_slots.CollectibleInventory;
 
 import java.io.IOException;
