+++ src/main/java/usecase_event/FightEvent.java
@@ -1,6 +1,6 @@
 package usecase_event;
 
-import entity.equipment_slots.item.Equipment;
+import entity.item.Equipment;
 import entity.monster.Monster;
 import entity.player.Player;
 import usecase_fight.DamageCalculator;
