+++ src/main/java/usecase_heal_and_upgrade/UpgradeCalculator.java
@@ -1,8 +1,7 @@
 package usecase_heal_and_upgrade;
 
-import entity.equipment_slots.item.Equipment;
+import entity.item.Equipment;
 import entity.player.Player;
-import usecase_heal_and_upgrade.*;
 
 public class UpgradeCalculator {
