+++ src/test/java/entity_tests/ArmorTest.java
@@ -1,6 +1,6 @@
 package entity_tests;
 
-import entity.equipment_slots.item.Armor;
+import entity.item.Armor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
