+++ src/test/java/entity_tests/BasicEquipmentSlotsTest.java
@@ -1,8 +1,8 @@
 package entity_tests;
 
-import entity.equipment_slots.item.Armor;
+import entity.item.Armor;
 import entity.equipment_slots.BasicEquipmentSlots;
-import entity.equipment_slots.item.Weapon;
+import entity.item.Weapon;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
