+++ src/test/java/entity_tests/WeaponTest.java
@@ -1,6 +1,6 @@
 package entity_tests;
 
-import entity.equipment_slots.item.Weapon;
+import entity.item.Weapon;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
